+++ src/senaite/fhir/catalog/indexer.py
@@ -3,6 +3,9 @@
 import json
 from plone.indexer import indexer
 from Products.CMFCore.interfaces import IContentish
+from senaite.core.interfaces import IContact
+from senaite.core.interfaces.catalog import IContactCatalog
+from senaite.fhir.behaviors.contact import getExternalID
 from senaite.fhir.interfaces import IFHIRCatalog
 from senaite.fhir import api as fapi
 
@@ -24,3 +27,11 @@ def fhir_resource_types(obj):
     """
     uids = fapi.get_fhir_uids(obj) or {}
     return json.dumps(uids)
+
+
+@indexer(IContact, IContactCatalog)
+def contact_external_id(instance):
+    """Indexes the external id assigned by the FHIR API consumer, so
+    Practitioner resources can be matched to an existing Contact by it
+    """
+    return getExternalID(instance) or ""