+++ src/senaite/fhir/setuphandlers.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+
+from bika.lims import api
+from senaite.core.catalog import CONTACT_CATALOG
 from senaite.core.setuphandlers import setup_core_catalogs
 from senaite.core.setuphandlers import setup_other_catalogs
 from senaite.fhir import logger
@@ -12,10 +15,18 @@ CATALOGS = (
 
 # Tuples of (catalog, index_name, index_attribute, index_type)
 INDEXES = [
+    (CONTACT_CATALOG, "contact_external_id", "", "FieldIndex"),
 ]
 
 # Tuples of (catalog, column_name)
 COLUMNS = [
+]
+
+# Tuples of (portal_type, list of behaviors)
+BEHAVIORS = [
+    ("Contact", [
+        "senaite.fhir.contact.IExtendedContactBehavior"
+    ]),
 ]
 
 
@@ -30,6 +41,9 @@ def setup_handler(context):
 
     # Setup catalogs
     setup_catalogs(portal)
+
+    # Add behaviors
+    setup_behaviors(portal)
 
     logger.info("{} setup handler [DONE]".format(PRODUCT_NAME.upper()))
 
@@ -72,6 +86,9 @@ def post_uninstall(portal_setup):
     context = portal_setup._getImportContext(profile_id)  # noqa
     portal = context.getSite()  # noqa
 
+    # Remove additional behaviors
+    remove_behaviors(portal)
+
     logger.info("{} uninstall handler [DONE]".format(PRODUCT_NAME.upper()))
 
 
@@ -80,3 +97,37 @@ def setup_catalogs(portal):
     """
     setup_core_catalogs(portal, catalog_classes=CATALOGS)
     setup_other_catalogs(portal, indexes=INDEXES, columns=COLUMNS)
+
+
+def setup_behaviors(portal):
+    """Assigns additional behaviors to existing content types
+    """
+    logger.info("Setup Behaviors ...")
+    pt = api.get_tool("portal_types")
+    for portal_type, behavior_ids in BEHAVIORS:
+        fti = pt.get(portal_type)
+        if not hasattr(fti, "behaviors"):
+            # Skip, type is not registered yet probably (AT2DX migration)
+            logger.warn("Behaviors is missing: {} [SKIP]".format(portal_type))
+            continue
+        fti_behaviors = fti.behaviors
+        additional = filter(lambda b: b not in fti_behaviors, behavior_ids)
+        if additional:
+            fti_behaviors = list(fti_behaviors)
+            fti_behaviors.extend(additional)
+            fti.behaviors = tuple(fti_behaviors)
+
+    logger.info("Setup Behaviors [DONE]")
+
+
+def remove_behaviors(portal):
+    """Remove behaviors added by this add-on to existing content types
+    """
+    logger.info("Removing Behaviors ...")
+    pt = api.get_tool("portal_types")
+    for portal_type, behavior_ids in BEHAVIORS:
+        fti = pt.get(portal_type)
+        orig_behaviors = filter(lambda b: b not in behavior_ids, fti.behaviors)
+        fti.behaviors = tuple(orig_behaviors)
+
+    logger.info("Removing Behaviors [DONE]")